Refactor vulnix logic from cli
